2nd time system ok

+++ backend/src/services/customer.service.js
@@ -2,6 +2,7 @@
 import prisma from "../db/prisma.js";
 import pkg from "@prisma/client";
 import addContainsToWhere from "../utils/addConstrainsToWhere.js";
+import AppError from "../utils/error.util.js";
 const { Decimal } = pkg;
 
 export const getCustomers = async (where, { page, limit }) => {
@@ -89,5 +90,16 @@ export const updateCustomer = async (data) =>
         where: { id: data.id },
         data,
     });
-export const deleteCustomer = async (id) =>
-    await prisma.customer.delete({ where: { id } });
+
+export const deleteCustomer = async (id, next) => {
+    const hasSales = await prisma.sale.count({
+        where: { customerId: id },
+    });
+
+    console.log(hasSales);
+
+    if (hasSales > 0)
+        throw new AppError("Cannot delete customer.\nWhen it is linked to one or more sales.", 409)
+
+    return await prisma.customer.delete({ where: { id } });
+}
